src/shapefile-tool/requirements.txt
flask>=3.0
geopandas>=1.0
shapely>=2.0
psycopg2-binary>=2.9
